Finished adding bloom - needs cleanup and some restructuring on the photo effects side, but this is currenly scoped out
