+++ OGL - EGE/engine/shader/BloomProcessor.hpp
@@ -15,63 +15,10 @@
 #include <list>
 #include "SceneManager.hpp"
 #include "Utilities.hpp"
-class HighResProcessor: public PostProcessor {
-public:
-    HighResProcessor(ShaderProgram * shader, Dimensions * d) : PostProcessor(shader, d) {
-        this->frame->bind();
-        this->addBuffer<HDRTextureBuffer>("texBuffer", new HDRTextureBuffer( d->width, d->height));
-        this->addBuffer<DepthBuffer>("depthBuffer", new DepthBuffer(d->width, d->height));
-        this->draw();
-        this->frame->unbind();
-    }
-};
-
-class LuminanceProcessor: public PostProcessor {
-    std::unordered_map<std::string, IGLObject*> * objects;
-public:
-    LuminanceProcessor(ShaderProgram * shader, Dimensions * d) : PostProcessor(shader, d) {
-        this->objects = new std::unordered_map<std::string, IGLObject*>();
-        this->objects->insert({"plane", new GLObject(RectangularPlane::getInstance())});
-        this->frame->bind();
-        this->addBuffer<LowResTextureBuffer>("texBuffer", new LowResTextureBuffer( d->width,  d->height));
-        this->addBuffer<DepthBuffer>("depthBuffer", new DepthBuffer(d->width, d->height));
-        this->draw();
-        this->frame->unbind();
-    }
-    
-    std::unordered_map<std::string, StorageBuffer*> * render(Scene * scene, Dimensions * viewPort, std::unordered_map<std::string, IGLObject*> * objects) {
-        return  PostProcessor::render(scene, viewPort, this->objects);
-    }
-};
-
-class PostPorcessorPipeline {
-public:
-    
-    std::list<PostProcessor*> * processors;
-    
-    PostProcessor* getLast() {
-        return this->processors->back();
-    }
-
-};
-
-class GaussianProcessor: public PostProcessor {
-    std::unordered_map<std::string, IGLObject*> * objects;
-public:
-    GaussianProcessor(ShaderProgram * shader, Dimensions * d) : PostProcessor(shader, d) {
-        this->objects = new std::unordered_map<std::string, IGLObject*>();
-        this->objects->insert({"plane", new GLObject(RectangularPlane::getInstance())});
-        this->frame->bind();
-        this->addBuffer<HDRTextureBuffer>("texBuffer", new HDRTextureBuffer( d->width,  d->height));
-        this->addBuffer<DepthBuffer>("depthBuffer", new DepthBuffer(d->width, d->height));
-        this->draw();
-        this->frame->unbind();
-    }
-    
-    std::unordered_map<std::string, StorageBuffer*> * render(Scene * scene, Dimensions * viewPort, std::unordered_map<std::string, IGLObject*> * objects) {
-        return  PostProcessor::render(scene, viewPort, this->objects);
-    }
-};
+#include "HDRProcessor.hpp"
+#include "LuminanceProcessor.hpp"
+#include "PostProcessorPipeline.hpp"
+#include "GaussianProcessor.hpp"
 
 class BloomBlenderProcessor: public PostProcessor {
     std::unordered_map<std::string, IGLObject*> * objects;
@@ -97,16 +44,18 @@ public:
     BloomProcessor(Dimensions * d) {
         this->processors = new std::list<PostProcessor*>();
         std::list<PostProcessor*>::iterator it = this->processors->begin();
+        // HIGH RES RENDER OF SCENE
         HighResProcessor * highRes = new HighResProcessor(NULL, d);
         this->processors->push_back(highRes);
         ShaderProgram * lumShader = ShaderManager::createShaderProgram("/Users/johnathontheriot/Desktop/OGL - EGE/OGL - EGE/luminance.vertex.glsl", "/Users/johnathontheriot/Desktop/OGL - EGE/OGL - EGE/luminance.fragment.glsl");
         lumShader->bindVars = [highRes](ShaderProgram * shader, GLObject* obj, Scene* scene) {
             shader->bindVariable<Texture>("tex", highRes->buffers->at("texBuffer"));
         };
-
+        // RENDER LUMINANCE
         LuminanceProcessor * lum = new LuminanceProcessor(lumShader, d);
         this->processors->push_back(lum);
         float * weights = Utilities::gaussianWeights(1, 5);
+        // BLUR LUMINANCE
         ShaderProgram * gaussShader = ShaderManager::createShaderProgram("/Users/johnathontheriot/Desktop/OGL - EGE/OGL - EGE/gaussian_blur.vertex.glsl", "/Users/johnathontheriot/Desktop/OGL - EGE/OGL - EGE/gaussian_blur.fragment.glsl");
         gaussShader->bindVars = [lum, weights](ShaderProgram * shader, GLObject* obj, Scene* scene) {
             shader->bindVariable<Texture>("tex", lum->buffers->at("texBuffer"));
@@ -126,7 +75,7 @@ public:
 
         GaussianProcessor * gaussSecond = new GaussianProcessor(gaussShaderSecond, d);
         this->processors->push_back(gaussSecond);
-        
+        // APPLY BLOOM
         ShaderProgram * bloomShader = ShaderManager::createShaderProgram("/Users/johnathontheriot/Desktop/OGL - EGE/OGL - EGE/bloom_blend.vertex.glsl", "/Users/johnathontheriot/Desktop/OGL - EGE/OGL - EGE/bloom_blend.fragment.glsl");
         bloomShader->bindVars = [gaussSecond, highRes](ShaderProgram * shader, GLObject* obj, Scene* scene) {
             shader->bindVariable<Texture>("scene", highRes->buffers->at("texBuffer"));
