+++ OGL - EGE/luminance.fragment.glsl
@@ -9,7 +9,7 @@ float luminance(vec3 rgb) {
 }
 
 void main(){
-    float threshold = 0.6;
+    float threshold = 0.8;
     color = texture(tex, vec2(vUV.r, vUV.g));
     if(luminance(color.rgb) <= threshold) {
         color = vec4(0, 0, 0, 0);
